+++ src/Loader/mmdBufferKind.ts
@@ -11,5 +11,9 @@ export class MmdBufferKind {
     public static readonly MatricesSdefRW0Kind = "matricesSdefRW0";
     public static readonly MatricesSdefRW1Kind = "matricesSdefRW1";
 
+    // for prevent data loss
+    public static readonly MatricesSdefR0Kind = "matricesSdefR0";
+    public static readonly MatricesSdefR1Kind = "matricesSdefR1";
+
     public static readonly EdgeScaleKind = "edgeScale";
 }